JPA repository test + add method existsCustomerWithId

+++ spring-boot-example/src/main/java/com/mihailcornescu/customer/CustomerService.java
@@ -40,6 +40,9 @@ public class CustomerService {
     }
 
     public void deleteCustomerById(Long id) {
+        if (!customerDao.existsCustomerWithId(id)) {
+            throw new ResourceNotFoundException("customer with id [%s] was not found.".formatted(id));
+        }
         customerDao.deleteCustomerById(id);
     }
